Add *.pem to .gitignore
Prevent uploading key files to remote repositories at least when
the .pem file extension is used for that.

Quarkus application of this project uses JWT as authentication and
authorization method. A private and public key pair can be used to
sign and verify the tokens and that key files could be placed in
the filesystem depending on configuration.