Create a common layout component
Layout component is used as a wrapper of components in each page
definition components.

Redux Toolkit is used to manage the global layout's state.
- Create a store using configureStore from @reduxjs/toolkit
- Create a slice using createSlice from @reduxjs/toolkit

A temporary dummy page implemented the layout component is
rendered at the /initial-page path. The root path / is set to be
redirected to the path by React Router.

Add dependencies to suppress a warning during executing npm start:
@babel/plugin-proposal-private-property-in-object